Descriptor cloning & copying

+++ src/engine/rendering/types.hpp
@@ -8,8 +8,8 @@
 
 namespace fgl::engine
 {
-	using PresentIndex = std::uint16_t;
-	using FrameIndex = std::uint16_t;
+	using PresentIndex = std::uint8_t;
+	using FrameIndex = std::uint8_t;
 
 	using CameraIndex = std::uint16_t;
